Chore: Moving document extensions to extensions.js
